Updated package "platform" to version 1.0.13.  Separated the identifiers "macosx*-i386-x86_64" and "macosx-universal" as incompatible with each other.
Plus addition of a cast in tclIO.c to match types in a comparison
which otherwise bugs a Windows debug build (warnings are errors).

+++ generic/tclIO.c
@@ -5405,7 +5405,7 @@ ReadChars(
      * for sizing receiving buffers.
      */
 
-    int toRead = ((unsigned) charsToRead > srcLen) ? srcLen : charsToRead;
+    int toRead = ((unsigned) charsToRead > (unsigned) srcLen) ? srcLen : charsToRead;
 
     /*
      * 'factor' is how much we guess that the bytes in the source buffer will
